+++ src/frontend/js/utils/test/factories/joanie.ts
@@ -118,6 +118,7 @@ export const TargetCourseFactory = factory((): TargetCourse => {
     course_runs: courseRuns,
     is_graded: true,
     position: 1,
+    is_self_paced: false,
   };
 });
 
@@ -274,6 +275,7 @@ export const CourseFactory = factory((): CourseListItem => {
     course_run_ids: [faker.string.uuid()],
     state: CourseStateFactory().one(),
     cover: JoanieFileFactory().one(),
+    is_self_paced: false,
   };
 });
 
@@ -283,6 +285,7 @@ export const CourseLightFactory = factory((): CourseLight => {
     code: faker.string.alphanumeric(5),
     title: FactoryHelper.sequence((counter) => `Course ${counter}`),
     cover: JoanieFileFactory().one(),
+    is_self_paced: false,
   };
 });
 
@@ -344,6 +347,7 @@ export const CourseListItemFactory = factory((): CourseListItem => {
     course_run_ids: [],
     state: CourseStateFactory().one(),
     cover: JoanieFileFactory().one(),
+    is_self_paced: false,
   };
 });