+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/SyllabusCourseRunCompacted/index.tsx
@@ -1,5 +1,4 @@
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
-import { Button } from '@openfun/cunningham-react';
 import { CourseRun, CourseRunDisplayMode } from 'types';
 import useDateFormat from 'hooks/useDateFormat';
 import { extractResourceId, isJoanieResourceLinkProduct } from 'api/lms/joanie';
@@ -17,29 +16,19 @@ const messages = defineMessages({
     description: 'Title of the enrollment dates section of an opened course run block',
     defaultMessage: 'Enrollment',
   },
-  course: {
-    id: 'components.SyllabusCourseRun.course',
-    description: 'Title of the course dates section of an opened course run block',
-    defaultMessage: 'Course',
-  },
   languages: {
     id: 'components.SyllabusCourseRun.languages',
     description: 'Title of the languages section of an opened course run block',
     defaultMessage: 'Languages',
   },
-  runPeriod: {
-    id: 'components.SyllabusCourseRun.enrollmentPeriod',
-    description: 'Enrollment date of an opened course run block',
-    defaultMessage: 'From {startDate} {endDate, select, undefined {} other {to {endDate}}}',
-  },
-  coursePeriod: {
-    id: 'components.SyllabusCourseRun.coursePeriod',
-    description: 'Course date of an opened course run block',
-    defaultMessage: 'From {startDate} {endDate, select, undefined {} other {to {endDate}}}',
+  selfPaceRunPeriod: {
+    id: 'components.SyllabusCourseRun.selfPaceEnrollmentPeriod',
+    description: 'Enrollment date of an opened and self paced course run block',
+    defaultMessage: 'Available until {endDate}',
   },
 });
 
-const OpenedCourseRun = ({
+const OpenedSelfPacedCourseRun = ({
   courseRun,
   showLanguages,
 }: {
@@ -48,37 +37,21 @@ const OpenedCourseRun = ({
 }) => {
   const formatDate = useDateFormat();
   const intl = useIntl();
-  const enrollmentStart = courseRun.enrollment_start
-    ? formatDate(courseRun.enrollment_start)
-    : '...';
   const enrollmentEnd = courseRun.enrollment_end ? formatDate(courseRun.enrollment_end) : '...';
-  const start = courseRun.start ? formatDate(courseRun.start) : '...';
-  const end = courseRun.end ? formatDate(courseRun.end) : '...';
   return (
     <>
       {courseRun.title && <h3>{StringHelper.capitalizeFirst(courseRun.title)}</h3>}
       <dl>
-        <dt>
-          <FormattedMessage {...messages.enrollment} />
-        </dt>
+        {!showLanguages && (
+          <dt>
+            <FormattedMessage {...messages.enrollment} />
+          </dt>
+        )}
         <dd>
           <FormattedMessage
-            {...messages.runPeriod}
+            {...messages.selfPaceRunPeriod}
             values={{
-              startDate: enrollmentStart,
               endDate: enrollmentEnd,
-            }}
-          />
-        </dd>
-        <dt>
-          <FormattedMessage {...messages.course} />
-        </dt>
-        <dd>
-          <FormattedMessage
-            {...messages.runPeriod}
-            values={{
-              startDate: start,
-              endDate: end,
             }}
           />
         </dd>
@@ -94,15 +67,15 @@ const OpenedCourseRun = ({
       {findLmsBackend(courseRun.resource_link) ? (
         <CourseRunEnrollment courseRun={courseRun} />
       ) : (
-        <Button className="course-run-enrollment__cta" href={courseRun.resource_link} fullWidth>
+        <a className="course-run-enrollment__cta" href={courseRun.resource_link}>
           {StringHelper.capitalizeFirst(courseRun.state.call_to_action)}
-        </Button>
+        </a>
       )}
     </>
   );
 };
 
-export const SyllabusCourseRun = ({
+export const SyllabusCourseRunCompacted = ({
   courseRun,
   course,
   showLanguages,
@@ -121,7 +94,7 @@ export const SyllabusCourseRun = ({
             compact={courseRun.display_mode === CourseRunDisplayMode.COMPACT}
           />
         ) : (
-          <OpenedCourseRun courseRun={courseRun} showLanguages={showLanguages} />
+          <OpenedSelfPacedCourseRun courseRun={courseRun} showLanguages={showLanguages} />
         )}
       </div>
     </DjangoCMSTemplate>